1、对新API接口进行问题调试； 2、去除老API相关代码

+++ Swing/M/CoreData/DBHelper.h
@@ -16,7 +16,7 @@
 + (void)clearDatabase;
 
 + (BOOL)addEvent:(EventModel*)model;
-+ (BOOL)delEvent:(int)objId;
++ (BOOL)delEvent:(int64_t)objId;
 
 + (BOOL)addTodo:(ToDoModel*)model;
