+++ Swing/M/KidModel.h
@@ -10,7 +10,7 @@
 
 @interface KidModel : JSONModel
 
-@property (nonatomic) int objId;
+@property (nonatomic) int64_t objId;
 @property (strong, nonatomic) NSString* firstName;
 @property (strong, nonatomic) NSString* lastName;
 @property (strong, nonatomic) NSString<Optional>* nickName;
